feat/#9: category 폴더 구조 수정
