+++ src/app/layouts/CategoryLayout.tsx
@@ -1,10 +1,12 @@
 import { Outlet } from 'react-router-dom';
+import BannerCarousel from '@/widgets/BannerCarousel';
 import CategoryTabs from '@/widgets/CategoryTabs';
 
 const CategoryLayout = () => {
   return (
     <>
       <CategoryTabs />
+      <BannerCarousel />
       <Outlet />
     </>
   );
